+++ packages/web-components/@sfx/base/src/utils.ts
@@ -1,7 +1,6 @@
 /*
- * The below functions were pulled
- * directly from
- * github.com/Polymer/lit-element/issues/42#issuecomment-442894676.
+ * The below functions were pulled directly from
+ * https://github.com/Polymer/lit-element/issues/42#issuecomment-442894676
  */
 
 /* istanbul ignore file */