+++ packages/web-components/@sfx/ui/src/list.ts
@@ -9,18 +9,20 @@ export default class List extends Base {
   /**
    * The text used in the header.
    */
-  @property({ type: String, reflect: true }) sfxListTitle: string = '';
+  @property({ type: String, reflect: true }) caption: string = '';
   /**
    * List data used to generate list.
    */
-  @property({ type: Array }) items: Array<any> = [];
+  @property({ type: Array }) items: any[] = [];
+
+  private randomStringId = 
 
   render() {
     return html`
-      ${this.sfxListTitle
+      ${this.caption
         ? html`
-            <h4 id="list_title_${this.sfxListTitle}">${this.sfxListTitle}</h4>
-            <ul aria-labelledby="list_title_${this.sfxListTitle}">
+            <h4 id="list_title_${this.caption}">${this.caption}</h4>
+            <ul aria-labelledby="list_title_${this.caption}">
               ${this.items.map(
                 item =>
                   html`